+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Updated Research Project tweaks [#3112](https://github.com/rokwire/illinois-app/issues/3112).
 
 ## [4.3.39] - 2023-04-10
 ### Deleted